LayoutTests/http/tests/dom/window-open-about-blank-and-access-document-expected.txt

CONSOLE MESSAGE: Injected script running.
PASS newWindow.document is defined.
